requirements.txt
python-memcached
python-telegram-bot
requests
SQLAlchemy
pycrypto